refactor: shrink python cli package commands
